refactor(app): extract TelemetryBuilder to runtime/telemetry.py (#74)
Move battery_snapshot, _supervisor_snapshot, _health_snapshot, and
telemetry from Runtime into TelemetryBuilder (same pattern as
HardwareGlue #73, DepthService #71, DeviceManager #70). Runtime keeps
shims for all 4 methods so all call-sites (ui/server.py, tests) are
unaffected. Behaviour-preserving move only; no logic changes.
